requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
redis>=5.0.0
litellm>=1.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-telegram-bot>=20.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
